Support multiple galleries
